requirements.txt
google-genai
google-generativeai
langchain-core
langchain-google-genai
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
numpy
opencv-python
openpyxl
pandas
pdfplumber
pillow
pydantic
PyMuPDF
pytesseract
PyYAML
reportlab
requests